+++ tests/Generator/Fixtures/ObjectWithoutProps/schema.yaml
