+++ tests/Generator/Fixtures/ObjectWithoutPropsUnion/Output/Foo.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Ns\EmptyNestedObject;
+namespace Ns\ObjectWithoutPropsUnion;
 
 class Foo
 {
@@ -14,7 +14,10 @@ class Foo
     private static array $schema = [
         'properties' => [
             'foo' => [
-                'type' => 'object',
+                'type' => [
+                    'string',
+                    'object',
+                ],
             ],
             'bar' => [
                 'type' => 'string',
@@ -26,7 +29,7 @@ class Foo
     ];
 
     /**
-     * @var mixed
+     * @var string|mixed
      */
     private $foo;
 
@@ -36,7 +39,7 @@ class Foo
     private ?string $bar = null;
 
     /**
-     * @param mixed $foo
+     * @param string|mixed $foo
      */
     public function __construct($foo)
     {
@@ -44,7 +47,7 @@ class Foo
     }
 
     /**
-     * @return mixed
+     * @return string|mixed
      */
     public function getFoo()
     {
@@ -60,17 +63,11 @@ class Foo
     }
 
     /**
-     * @param mixed $foo
+     * @param string|mixed $foo
      * @return self
      */
     public function withFoo($foo) : self
     {
-        $validator = new \JsonSchema\Validator();
-        $validator->validate($foo, self::$schema['properties']['foo']);
-        if (!$validator->isValid()) {
-            throw new \InvalidArgumentException($validator->getErrors()[0]['message']);
-        }
-
         $clone = clone $this;
         $clone->foo = $foo;
 
@@ -121,7 +118,10 @@ class Foo
             static::validateInput($input);
         }
 
-        $foo = $input->{'foo'};
+        $foo = match (true) {
+            is_string($input->{'foo'}), true => $input->{'foo'},
+            default => throw new \InvalidArgumentException("could not build property 'foo' from JSON"),
+        };
         $bar = isset($input->{'bar'}) ? $input->{'bar'} : null;
 
         $obj = new self($foo);
@@ -137,7 +137,9 @@ class Foo
     public function toArray() : array
     {
         $output = [];
-        $output['foo'] = $this->foo;
+        $output['foo'] = match (true) {
+            is_string($this->foo), true => $this->foo,
+        };
         if (isset($this->bar)) {
             $output['bar'] = $this->bar;
         }
@@ -168,5 +170,11 @@ class Foo
 
         return $validator->isValid();
     }
-}
 
+    public function __clone()
+    {
+        $this->foo = match (true) {
+            is_string($this->foo), true => $this->foo,
+        };
+    }
+}
