+++ tests/Generator/Fixtures/ObjectWithoutPropsUnion/schema.yaml
